requirements.txt
requests==2.31.0
pytest==7.4.4
allure-pytest==2.13.2
assertpy==1.1
pytest-ordering==0.6
allure-python-commons==2.13.2
